device-mapper: Do not override error code from dm_get_device()
Some of the device mapper targets right now override return code returned
by dm_get_device() and return either -EINVAL or -ENXIO. It is not clear
what do we gain by that. It is better to propagate return code as it is
to caller.

I specifically ran into one issue where underlying device was busy but
I kept on getting -EINVAL. After this change we will get -EBUSY instead
and it will be easier to figure out what's the problem.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>