4338-dhowells-selinux-handle-opening-of-a-unioned-file
To: linux-unionfs@vger.kernel.org, selinux@tycho.nsa.gov
Cc: mjg59@srcf.ucam.org, dwalsh@redhat.com, linux-kernel@vger.kernel.org,
 eparis@redhat.com, dhowells@redhat.com,
 linux-security-module@vger.kernel.org, linux-fsdevel@vger.kernel.org,
 sds@tycho.nsa.gov
Subject: [PATCH 4/5] SELinux: Handle opening of a unioned file
Organization: Red Hat UK Ltd. Registered Address: Red Hat UK Ltd, Amberley
 Place, 107-111 Peascod Street, Windsor, Berkshire, SI4 1TE, United Kingdom.
 Registered in England and Wales under Company Registration No. 3798903
User-Agent: StGit/0.17.1-dirty
Message-ID: <20150928200102.8141.53879.stgit@warthog.procyon.org.uk>

Tracking: 4338-dhowells-selinux-handle-opening-of-a-unioned-file.patch
CVE:
Git:
Bugzilla:
Status: unresolved
Status-tech: need acks;
Status-management: unresolved
Series-info:
Test-build-status:
------------------------------------------------------------------------

Handle the opening of a unioned file by trying to derive the label that would
be attached to the union-layer inode if it doesn't exist.

If the union-layer inode does exist (as it necessarily does in overlayfs, but
not in unionmount), we assume that it has the right label and use that.
Otherwise we try to get it from the superblock.

If the superblock has a globally-applied label, we use that, otherwise we try
to transition to an appropriate label.  This union label is then stored in the
file_security_struct.

We then perform an additional check to make sure that the calling task is
granted permission by the union-layer inode label to open the file in addition
to a check to make sure that the task is granted permission to open the lower
file with the lower inode label.

Signed-off-by: David Howells <dhowells@redhat.com>