ovl: Do not leave whiteout during unlink/rmdir if parent does have not merge property
I am trying to fix following bug
https://bugzilla.kernel.org/show_bug.cgi?id=109611

Do following.

$ mkdir upper lower work merged upper/dir/
$ touch lower/test
$ sudo mount -t overlay overlay -olowerdir=lower,upperdir=upper,workdir=work merged
$ mv merged/test merged/dir/
$ rm merged/dir/test
$ ls -l merged/dir/
/usr/bin/ls: cannot access merged/dir/test: No such file or directory
total 0
c????????? ? ? ? ?            ? test

Basic problem seems to be that once a file has been unlinked, a
whiteout has been left behind and doing ls makes that whiteout visible.

whiteout is visible because parent dir is of not type MERGE, hence od->is_real
is set during ovl_dir_open(). And that means ovl_iterate() passes on iterate
handling directly to underlying fs. Underlying fs does not know/filter
whiteouts so it becomes visible to user.

Why did we leave a whiteout to begin with when we should not have.
ovl_do_remove() checks for OVL_TYPE_PURE_UPPER() and does not leave
whiteout if file is pure upper. In this case file is not found to be
pure upper hence whiteout is left.

So why file was not PURE_UPPER in this case? I think because dentry is
still carrying some leftover state which was valid before rename. For example,
od->numlower was set to 1 as it was a lower file. After rename, this state
is not valid anymore as there is no such file in lower.

Also during rename, we first copied up the file and set oe->opaque=1. That
means this file will never be characterized as PURE.

And hence we leave a whiteout when we should not have to. So I have put a
temporary hack to check if parent is type MERGE or not. If it is not then
looks like we might not have to leave a whiteout.

This seems like a hack fix to me. Real problem seems to be that during rename
some ovl_entry state got stale. Not sure how to take care of that. So
sending this hack fix out and hoping somebody can suggest what's a better
way to fix it.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>