overlayfs: Call security hooks on upper before unlink and rmdir operations
Right unlink and rmdir on upper don't call security hooks. Instead using
the creds of mounter, we do vfs_rename().

First call security hooks in the context of task to make sure task is
allowed to unlink/rmdir a file/dir on upper, and then continue with
vfs_rename() operation.

If file/dir is present only in upper, then we call vfs_unlink() or
vfs_rmdir() in task context and checks will be done by vfs anyway. So
we don't do these checks in that case.

If file is present only in lower, then I guess it should be ok to put
whiteout in upper without checks. We are not touching file on lower
anyway.

These additional checks are done only if file is present both in upper
and lower. In that case we perform vfs_rename() in the context of
mounter. So it should be a good idea to do checks first in the context
of task.
 
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>