overlayfs: Call security hooks during file creation time
When a file or dir is created on upper (regular, symlink, hardlink, mknod),
overlay can call respective vfs_* helper in the context of task on upper
filesystem, if no whiteout is present in upper/ dir. In that case vfs helpers
should call security hooks on upper fs.

But if a whiteout is present in upper/, in that case overlay creates file
in workdir and then does vfs_rename(file, whiteout). And security hooks
are not called on upper fs. 

Fix this by calling security hooks in the context of task when whiteout is
present on upper/. If checks pass, only then continue with switching
creds to mounter and doing vfs_rename().

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>