security, overlayfs: Correctly label a newly created file
When a file is newly crated in upper dir, it needs to be labeled
appropriately. 

For non-context moutns, it gets the label based on parent directory's label
but for context mounts it should get label as specified in context=
mount option.

This patch adds a new security hook which can prepare a new set of creds 
and switch to these creds. After creating file, caller should revert
back to old creds and that should release the creds prepared by hook.

This new set of creds should make sure that file is created with correct
label for context mounts.
 
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>