vfs, overlayfs, security: A hook to check inode permission on lower inode
When a file is opened through overlay,  inode_permission() is called
which does both DAC and MAC checks. DAC checks happen on real inode of
file (lower/upper) while MAC checks happen on overlay inode as well as
real inode.

All these checks happen against tasks's creds and if task is openeing file
for WRITE, then file should have WRITE access to task for open to succeed.

SELinux developers are planning to label files in such a way so that lower
files can be opended for READ by task but not for WRITE. This is to make
sure that even if container task manages to access file outside the overlay
mount, it can't WRITE to file on lower.

But this also means that SELinux check on lower inode fails when task
tries to open file for WRITE (as task does not have WRITE permission on
lower).

One argument has been that if file is opened for WRITE it will be copied
up anyway with a label accessible to task (in the case of context mount).
So doing a MAY_WRITE check on lower inode is not strictly necessary.

This patch creates a separate vfs function inode_permission_lower(), which
should be called when doing permission check on lower inode. Only difference
between __inode_permission() and inode_permission_lower() is that later
also takes in pointer to corresponding overlay inode as parameter and passes
it to security hook.

This patch also creates a new security hook for checking permission on
lower inode. SELinux implementation of hook resets MAY_WRITE and MAY_APPEND
from mask for the case of context mounts. For non-context mounts, MAC and
DAC checks are same and continue to use original mask for lower inode.

Finally, overlay calls vfs helper __inode_permission() for upper inode
and inode_permission_lower() for lower inode.

Reviewed-by: Stephen Smalley <sds@tycho.nsa.gov>
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>