overlayfs: Correctly label newly created file over whiteout
File creation in upper can take two paths. If there is no whiteout present
in upper dir, then file is directly created on upper in the context of
task. But if there is a whiteout present, then file is created in work/
dir and then renamed in upper dir, and mounter's creds are used for file
creation in work dir.

Even if file is created in work/ dir, it should get a label as if task
created it in upper.

This patch introduces a new hook which determines the label dentry will
get if it had been created by task in upper and sets the secid of label
in passed set of creds. Caller makes use of these new creds for file
creation.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>