ovl: Create origin xattr on copy up for all files
Right now my understanding is that origin xattr is created for all copied
up files if index=on. And if index=off, then we create it for all type
of files except hardlinks (nlink != 1).

With metadata only copy up, I will still require origin xattr to copy up
data later, so create it even for hardlinks even with index=off.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>