ovl: A new xattr OVL_XATTR_METACOPY for file on upper
Now we will have the capability to have upper inodes which might be only
metadata copy up and data is still on lower inode. So add a new xattr
OVL_XATTR_METACOPY to distinguish between two cases. 

Presence of OVL_XATTR_METACOPY reflects that file has been copied up metadata
only and and data will be copied up later from lower origin.
So this xattr is set when a metadata copy takes place and cleared when
data copy takes place.

We also use a bit in ovl_inode->flags to cache OVL_UPPERDATA which reflects
whether ovl inode has data or not (as opposed to metadata only copy up).

If a file is copied up metadata only and later when same file is opened
for WRITE, then data copy up takes place. We copy up data, remove METACOPY
xattr and then set the UPPERDATA flag in ovl_entry->flags. While all
these operations happen with oi->lock held, read side of oi->flags can be
lockless. That is another thread on another cpu can check if UPPERDATA
flag is set or not.

So this gives us an ordering requirement w.r.t UPPERDATA flag. That is, if
another cpu sees UPPERDATA flag set, then it should be guaranteed that
effects of data copy up and remove xattr operations are also visible.

For example.

	CPU1				CPU2
ovl_d_real()				acquire(oi->lock)
 ovl_open_maybe_copy_up()                ovl_copy_up_data()
  open_open_need_copy_up()		 vfs_removexattr()
   ovl_already_copied_up()
    ovl_dentry_needs_data_copy_up()	 ovl_set_flag(OVL_UPPERDATA)
     ovl_test_flag(OVL_UPPERDATA)       release(oi->lock)

Say CPU2 is copying up data and in the end sets UPPERDATA flag. But if
CPU1 perceives the effects of setting UPPERDATA flag but not the effects
of preceeding operations (ex. upper that is not fully copied up), it will be
a problem.

Hence this patch introduces smp_wmb() on setting UPPERDATA flag operation
and smp_rmb() on UPPERDATA flag test operation.

May be some other lock or barrier is already covering it. But I am not sure
what that is and is it obvious enough that we will not break it in future.

So hence trying to be safe here and introducing barriers explicitly for
UPPERDATA flag/bit.

Note, we now set OVL_UPPERDATA on all kind of dentires and check
OVL_UPPERDATA only where it makes sense. If a file is created on upper,
we set OVL_UPPERDATA. This can be accessed lockless in ovl_d_real()
path again. This time ordering dependency is w.r.t oe->__upperdata. We
need to make surr OVL_UPPERDATA is visible before oe->__upperdata is
visible. Otherwise following code path might try to copy up an upper
only file.

ovl_d_real()
 ovl_open_maybe_copy_up()
  open_open_need_copy_up()
   ovl_already_copied_up()
    ovl_dentry_needs_data_copy_up()
     ovl_test_flag(OVL_UPPERDATA)
         <-- OVL_UPPERDATA is not visible yet. copy up file.

I have not introduced any barrier to take care of this. There is already
a check ovl_should_check_upperdata() to make sure lower dentry is there
otherwise return false (even if OVL_UPPERDATA is not visible yet).

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>