ovl: Allocate bigger stack for origin
Right now there is only one origin for upper dentry and we allocate space
to hold only one ovl_path. But with metacopy support, it is possible that
lower origin is a metacopy itself and we will continue the search for
lower most data entry and install that too in lower stack so that this
entry can be used for copy up later.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>