ovl: Get and install lower most data dentry in stack
Now we want to support midlayer metacopy dentry as well. To support that
we need to find out lowest data dentry and keep track of that too so that
later it can be used for data copy up.

Introduce logic to traverse through metacopies and find and install the
lowest data ovl_path in stack.

This patch does it only for upper dentry at this point of time. Next patch
will introduce this logic for the cases when upper does not exist.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>