ovl: Initialize ovl_inode->redirect in ovl_get_inode()
All the inode/ovl_inode properties ideally should be initialized in 
ovl_get_inode(). Otherwise this can be a problem for the cases where
multiple dentries point to same inode (hardlinks, disconnected dentries).

As of now this is not a problem as redirects are used only for directories
which don't share inode. But soon I want to use redirects for regular files
also and there it will become an issue. 

Hence, move ->redirect initialization in ovl_get_inode().

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>