ovl: Set OVL_INDEX flag in ovl_get_inode()
All the inode/ovl_inode initialization should happen in ovl_get_inode().
This is especially useful when multiple dentries are pointing to same
inode and inode is already in the cash. In that case, we don't have to
initialize the OVL_INDEX. We don't even take ovl_inode->lock in ovl_lookup()
and that can run into issues for the case of shared inode.

So move OVL_INDEX flag setting in ovl_get_inode().

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>