ovl: Set redirect on upper inode when it is linked
When we create a hardlink to a metacopy upper file, first the redirect
on that inode. Path based lookup will not work with newly created link
and redirect will solve that issue.

Also use absolute redirect as two hardlinks could be in different directores
and relative redirect will not work.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>