ovl: Do not error if REDIRECT XATTR is missing
Currently we first call vfs_getxattr() to get size of REDIRECT xatttr and
then make another call vfs_getxattr() to read the xattr in buffer.

This is all part of ovl_lookup() and we do not have ovl_inode->lock. That
means it is possible that inode got copied up on another cpu and is not
a metacopy inode anymore. And that also means REDIRECT xattr got removed.
And that means that while first call to vfs_getxattr() succeeded, second
call can fail with -ENODATA. 

Do not error out if -ENODATA is received from second call. With metacopy
enabled, it is not an error case anymore.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>