ovl: Re-check redirect xattr during inode initialization
So far redirect could be placed on directories only and now it can be
placed on regular files as well. Also it could be completely removed 
when a metacopy copy up file's data is copied up. That means if a redirect
is present during ovl_lookup(), it could be gone by the time ovl_get_inode()
happens. 

Or it is possible that ovl_lookup() does not see a redirect and a rename
is taking place on a hard link and that places a redirect. And by the
time ovl_lookup() calls ovl_get_inode(), it sets ovl_inode->redirect = NULL
(Assume inode got flushed out of cache and was allocated new).

IOW, because we check and process redirect without locks in ovl_lookup(),
many possibilities open up for regular files. So for such cases, do not
use the redirect provided by the caller. Instead query it and install
in ovl_inode->redirect. 

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>