ovl: Verify a data dentry has been found for metacopy inode
If we find a upper metacopy inode, make sure we also found associated data
dentry in lower. Otherwise copy up operation later will fail.

There are two cases where this can happen. First case is that somehow
data file was removed from lower layer. Other case is that REDIRECT
xattr was removed due to copy up of file on another cpu (when inode is
shared between two dentries) and hence ovl_lookup() could not find the
correct dentry.

First case is an error while second case is not error. If file has been
copied up, then it does not matter if data dentry was found or not.

Redirect removal is protected using ovl_inode->lock and ovl_lookup() does
not have access to that lock. So to differentiate between these two 
cases, take appropriate inode lock in ovl_get_inode() and make sure a
data dentry has been found for metacopy inode. Otherwise lookup failed
and its an error.

For example, say two files are hardlinked, foo.txt and bar.txt. Say foo.txt
is renamed to foo-renamed.txt gets copied up metadata only. This will also
put a redirect "/foo.txt" on hardlnk  inode. Now assume foo-renamed.txt
is opened for write and is undergoing data copy up on one cpu and bar.txt
is under going ovl_lookup() on other cpu. Data copy up path will remove
REDIRECT and METACOPY xattr. It is possible that METACOPY xattr is
visible to ovl_lookup() but by the REDIRECT xattr was gone by the time.
That means no data dentry will be found but at the same time now inode
is not metacopy inode. So data dentry is not required. So this is not
error case. But if inode was still metacopy but data dentry was not found
this is error case. (May be due to underlying layer changed). Fix it by
returning -ESTALE.

If inode was found in cache, then take ovl_inode->lock before checking
status of inode. If inode has been allocated, then it is returned with
inode lock anyway and other threads will block on that lock, so no need
to take ovl_inode->lock.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>